% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-multiomics.R
\name{benchmarkROC}
\alias{benchmarkROC}
\title{run_rank}
\usage{
benchmarkROC(
  score_lists,
  GS_list,
  cts = "all",
  min_num = 30,
  kins = "all",
  sub_samp_frac = 0.8,
  ROC_obj = F,
  metric = "auroc",
  ...
)
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
