% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-multiomics.R
\name{.getGSvals}
\alias{.getGSvals}
\title{run_rank}
\usage{
.getGSvals(score_df, GSpos_list, GSneg_list, scale = F)
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
