% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-multiomics.R
\name{.ROC_sampler}
\alias{.ROC_sampler}
\title{run_rank}
\usage{
.ROC_sampler(
  GSpos,
  GSneg,
  Npert = 1000,
  set_size = "GSpos_size",
  ROC_obj = F,
  seed = 1,
  ...
)
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
