% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{load_meta}
\alias{load_meta}
\title{load_meta}
\usage{
load_meta()
}
\value{
data frame containing additional information
about the perturbation experiments
}
\description{
Loads meta data from the perturbation collection
published by Hernandez-Armenta with the Hijazi datasets.
}
\examples{
load_meta()
}
