% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{load_perturbData}
\alias{load_perturbData}
\title{load_perturbData}
\usage{
load_perturbData()
}
\value{
data frame containing logFC of different
perturbation experiments
}
\description{
Loads data from the perturbation collection
published by Hernandez-Armenta with the Hijazi datasets.
}
\examples{
load_perturbData()
}
