% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-perturbation.R
\name{prepareBench}
\alias{prepareBench}
\title{prepareBench}
\usage{
prepareBench(act, meta, rm_experiments = F, method_id = "method")
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
