% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarKIN-perturbation.R
\name{remove_bg}
\alias{remove_bg}
\title{remove_bg}
\usage{
remove_bg(mat, meta)
}
\arguments{
\item{mat}{Matrix with experiments as rows and kinases as columns.}

\item{meta}{Data frame containing sample (experiment) information and perturb (target) in each experiment}
}
\value{
Matrix after filtering out experiments where no activity for the target was inferred.
}
\description{
Identifies experiments where no activity could be inferred for any of the
target kinases and removes the experiment.
}
\examples{
 # Create random network and matrix
 set.seed(123)

 mat <- data.frame(experiment = c("exp1", "exp2"),
                   A = runif(2, min = -2, max = 2),
                   B = runif(2, min = -3, max = 2),
                   C = runif(2, min = -2, max = 2),
                   D = runif(2, min = -3, max = 2))

 meta <- data.frame(sample = c("exp1", "exp2"),
                    perturb = c("A", "F"))

# Remove experiment from background
mat_filtered <- remove_bg(mat = mat, meta = meta)

}
