% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-perturbation.R
\name{run_perturbBench}
\alias{run_perturbBench}
\title{run_perturbBench}
\usage{
run_perturbBench(
  act,
  meta,
  scale_data = T,
  rm_experiments = F,
  n_iter = 1000,
  method_id = "method",
  metric = "auroc",
  ...
)
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
