% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarKIN-rank.R
\name{run_phit}
\alias{run_phit}
\title{run_phit}
\usage{
run_phit(act, meta, k = 10)
}
\arguments{
\item{act}{Activity matrix with kinases as rows and experiments as columns.}

\item{meta}{Data frame containing sample (experiment) information and perturb (target) in each experiment.}

\item{k}{Number of top kinases to test in each experiment.}
}
\value{
Data frame containing the rank of each perturbed kinase based on its activity.
}
\description{
Calculates the percentage of the perturbed kinase(s) appearing among the top
k kinases based on its activty.
}
\examples{
# Create random meta and matrix
set.seed(321)

mat <- data.frame(exp1 = runif(5, min = -2, max = 2),
                  exp2 = runif(5, min = -3, max = 2))
rownames(mat) <- c("A", "B", "C", "D", "E")

meta <- data.frame(id = c("exp1", "exp2"),
                   target = c("E", "A"),
                   sign = c(1, -1))

# run benchmark
res <- run_phit(act = mat, meta = meta, k = 1)

}
