% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-rank.R
\name{run_rank}
\alias{run_rank}
\title{run_rank}
\usage{
run_rank(act, meta)
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
