% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-rokai.R
\name{run_zscore}
\alias{run_zscore}
\title{run_zscore}
\usage{
run_zscore(mat, network, minsize = 5)
}
\arguments{
\item{mat}{Matrix with phosphorylation sites as rows and experiments as columns.}

\item{network}{Data frame containing kinase-substrate interactions.}

\item{minsize}{Integer indicating the minimum number of targets per source.}
}
\value{
A long format tibble of the activity scores for each source (kinase)
across the samples.
}
\description{
Calculates kinase activities from phosphoproteomics data
using the Z-score as implemented in RoKAI.
}
\details{
RoKAI zscore algorithm re-implementation in R
}
\examples{
# Create random network and matrix
set.seed(123)
net <- data.frame(source = rep(c("A", "B", "C"), each = 5),
                  target = sample(rep(c("A", "B", "C", "D", "E"), each = 3)),
                  mor = 1) \%>\%
           dplyr::distinct()

mat <- data.frame(exp1 = runif(5, min = -2, max = 2))
rownames(mat) <- c("A", "B", "C", "D", "E")

# Activity estimation
run_zscore(mat = mat, network = net, minsize = 2)
}
