% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarKIN-perturbation.R
\name{scale_scores}
\alias{scale_scores}
\title{scale_scores}
\usage{
scale_scores(mat, scaling = "sd")
}
\arguments{
\item{mat}{Matrix with kinases as rows and experiments as columns.}

\item{scaling}{Scaling based on max or standard deviation.}
}
\value{
Scaled matrix with kinases as rows and experiments as columns.
}
\description{
Scales inferred kinase activity scores per experiment. Standardizes standard
deviation or maximum score across experiments.
}
\examples{
# Create random network and matrix
set.seed(123)

mat <- data.frame(exp1 = runif(5, min = -2, max = 2),
                  exp2 = runif(5, min = -3, max = 2))
rownames(mat) <- c("A", "B", "C", "D", "E")

# Scaling
scale_scores(mat = mat)

}
