% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-multiomics.R
\name{.getGSvals}
\alias{.getGSvals}
\title{.getGSvals}
\usage{
.getGSvals(score_df, GSpos_list, GSneg_list, scale = F)
}
\arguments{
\item{score_df}{A dataframe of scores with kinases as row names and samples/conditions as column names.}

\item{GSpos_list}{A named list of positive gene sets for each kinase.}

\item{GSneg_list}{A named list of negative gene sets for each kinase.}

\item{scale}{A logical indicating whether to scale the scores by z-scores across each row. Default is FALSE.}
}
\value{
A list with the following elements:
\item{Z-scores}{Dataframe of (optionally scaled) scores.}
\item{GSpos_values}{List of extracted positive gene set values for each kinase.}
\item{GSneg_values}{List of extracted negative gene set values for each kinase.}
\item{number_GSpos_per_kinase}{Numeric vector of counts of positive gene sets per kinase.}
\item{number_GSneg_per_kinase}{Numeric vector of counts of negative gene sets per kinase.}
}
\description{
This function processes a dataframe of kinase scores, optionally scales them,
and extracts values based on specified positive and negative gene sets.
}
