% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{load_meta}
\alias{load_meta}
\title{load_meta}
\usage{
load_meta(dataset = "both", targets_hijazi = "manual")
}
\arguments{
\item{dataset}{whether to load the "hijazi", "hernandez" or "both" datasets}

\item{targets_hijazi}{whether to use a "manual" list of targets for the
inhibitors or whether to consider all kinases targets based on the "selectivity"}
}
\value{
data frame containing additional information
about the perturbation experiments
}
\description{
Loads meta data from the perturbation collection
published by Hernandez-Armenta with the Hijazi datasets.
}
\examples{
load_meta()
}
