% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-rank.R
\name{run_rank}
\alias{run_rank}
\title{run_rank}
\usage{
run_rank(act, meta)
}
\arguments{
\item{act}{Activity matrix with kinases as rows and experiments as columns.}

\item{meta}{Data frame containing sample (experiment) information and perturb (target) in each experiment.}
}
\value{
Data frame containing the rank of each perturbed kinase based on its activity.
}
\description{
Calculates the rank of the perturbed kinases in the respective experiment
}
\examples{
# Create random meta and matrix
set.seed(321)

mat <- data.frame(exp1 = runif(5, min = -2, max = 2),
                  exp2 = runif(5, min = -3, max = 2))
rownames(mat) <- c("A", "B", "C", "D", "E")

meta <- data.frame(id = c("exp1", "exp2"),
                   target = c("E", "A"),
                   sign = c(1, -1))

# run benchmark
res <- run_rank(act = mat, meta = meta)

}
