% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkeR-perturbation.R
\name{scale_scores}
\alias{scale_scores}
\title{scale_scores}
\usage{
scale_scores(mat, scaling = "sd")
}
\value{
data.frame with phosphorylation sites as rownames and perturbation
experiments as columns.
}
\description{
This function returns a matrix of logFC of phosphorylation sites
from perturbation experiments. These can be used to infer kinase
activities.
}
